+++ src/Infrastructure/OpenId/User/Statefull/StatefullOpenIdUserAppOptions.ts
@@ -1,0 +1,7 @@
+import type { OpenIdUserAppOptions } from '../OpenIdUserAppOptions';
+
+export interface StatefullOpenIdUserAppOptions extends OpenIdUserAppOptions {
+  session?: {
+    lifetime: string
+  }
+}