+++ src/Infrastructure/OpenId/index.ts
@@ -1,4 +1,2 @@
-export * from './OpenIdUserModule';
-export * from './Statefull';
-export * from './Stateless';
-export * from './Constants';
+export * from './User';
+export * from './OpenIdClientFactory';